ыиджет и взаимодействие с ним.

+++ data/weather-remote/src/main/kotlin/dmitry/molchanov/weather_remote/entity/Forecast.kt
@@ -8,7 +8,7 @@ data class Forecast (
     @SerialName("date")
     val date: String,
     @SerialName("date_ts")
-    val dateTs: Int? = null,
+    val dateTs: Long? = null,
     @SerialName("week")
     val week: Int,
     @SerialName("sunrise")
